Move SHA256 to Data.cs - Addressing other comments
